Add files via upload
This project is related to the forecast model for two future years based on data from three years. It has two files one PowerPoint file presented the comparison between two sectors and the second one is related to the forecast model for the companies and interact dashboard in the excel file.